# server/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: VIVIM Capture API
  description: |
    API for capturing AI conversations from various providers.

    ## Features

    * Capture conversations from Claude, ChatGPT, Gemini, Grok, DeepSeek, Kimi, Qwen, and Zai
    * Automatic provider detection
    * Rich content formatting with code blocks and diagrams
    * Conversation storage and retrieval
    * Analytics and statistics

    ## Authentication

    Currently in development. Authentication will be added in a future version.

  version: 2.0.0
  contact:
    name: VIVIM API Support
    url: https://github.com/vivim-ai/vivim-server
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.vivim.io/v1
    description: Production server
...
paths:
...
  /api/v1/detect-provider:
    get:
      tags: [Providers]
      summary: Detect provider from URL
      description: Detects which AI provider a URL belongs to
      operationId: detectProvider
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: The URL to analyze
          example: https://claude.ai/share/abc123
      responses:
        '200':
          description: Provider detected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectProviderResponse'
        '400':
          description: Invalid URL parameter
  /api/v1/capture:
    post:
      tags: [Capture]
      summary: Capture a conversation
      description: |
        Captures a conversation from an AI provider URL and stores it in the database.

        The API will automatically detect the provider from the URL, or you can specify it explicitly.

        Caching is enabled by default - recent captures (within 60 minutes) will be returned from cache.
      operationId: captureConversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaptureRequest'
            examples:
              claude:
                summary: Claude conversation
                value:
                  url: https://claude.ai/share/abc123def456
              chatgpt:
                summary: ChatGPT conversation
                value:
                  url: https://chat.openai.com/share/xyz789
              with_options:
                summary: With custom options
                value:
                  url: https://claude.ai/share/abc123
                  options:
                    timeout: 180000
                    richFormatting: true
                    metadataOnly: false
                    cache: true
                    cacheMinutes: 30
      responses:
        '200':
          description: Conversation captured successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureResponse'
              examples:
                success:
                  summary: New capture
                  value:
                    status: success
                    cached: false
                    data:
                      id: 550e8400-e29b-41d4-a716-446655440000
                      provider: claude
                      sourceUrl: https://claude.ai/share/abc123
                      title: 'Conversation about TypeScript'
                      createdAt: '2025-01-24T10:30:00Z'
                      exportedAt: '2025-01-24T10:35:00Z'
                      messageCount: 42
                      messages: [...]
                cached:
                  summary: Cached response
                  value:
                    status: success
                    cached: true
                    data:
                      id: 550e8400-e29b-41d4-a716-446655440000
                      provider: claude
                      sourceUrl: https://claude.ai/share/abc123
                      title: 'Conversation about TypeScript'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CaptureRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: The URL of the conversation to capture
          example: https://claude.ai/share/abc123
        options:
          type: object
          properties:
            timeout:
              type: integer
              minimum: 1000
              maximum: 300000
              default: 120000
              description: Capture timeout in milliseconds
            richFormatting:
              type: boolean
              default: true
              description: Enable rich content formatting (code blocks, diagrams)
            metadataOnly:
              type: boolean
              default: false
              description: Only return metadata, not full messages
            headless:
              type: boolean
              default: true
              description: Run browser in headless mode
            provider:
              type: string
              enum: [claude, chatgpt, gemini, grok, deepseek, kimi, qwen, zai]
              description: Force specific provider (auto-detected if not specified)
            cache:
              type: boolean
              default: true
              description: Use cache for recent captures
            cacheMinutes:
              type: integer
              minimum: 1
              default: 60
              description: Cache duration in minutes
    CaptureResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        cached:
          type: boolean
          description: True if returned from cache
          example: false
        data:
          $ref: '#/components/schemas/Conversation'
    Conversation:
      type: object
      required: [id, provider, sourceUrl, title, createdAt, exportedAt, messageCount]
      properties:
        id:
          type: string
          format: uuid
          description: Unique conversation identifier
        provider:
          type: string
          enum: [claude, chatgpt, gemini, grok, deepseek, kimi, qwen, zai]
          description: AI provider
        sourceUrl:
          type: string
          format: uri
          description: Original URL of the conversation
        title:
          type: string
          description: Conversation title
        createdAt:
          type: string
          format: date-time
          description: When the conversation was created
        exportedAt:
          type: string
          format: date-time
          description: When the conversation was exported
        messageCount:
          type: integer
          description: Number of messages in the conversation
        metadata:
          type: object
          properties:
            provider:
              type: string
            model:
              type: string
              description: AI model used (if available)
        messages:
          type: array
          description: Array of message objects
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              role:
                type: string
                enum: [user, assistant, system]
              content:
                oneOf:
                  - type: string
                    description: Plain text content
                  - type: array
                    description: Rich content blocks
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          enum: [text, code, mermaid, image]
                        content:
                          type: string
                        language:
                          type: string
                          description: Programming language (for code blocks)
              timestamp:
                type: string
                format: date-time
                nullable: true
...
    DetectProviderResponse:
      type: object
      required: [provider, supported]
      properties:
        provider:
          type: string
          nullable: true
          description: Detected provider (null if unknown)
          enum: [claude, chatgpt, gemini, grok, deepseek, kimi, qwen, zai, other, null]
        supported:
          type: boolean
          description: Whether the provider is supported
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Error code
              example: VALIDATION_ERROR
            message:
              type: string
              description: Human-readable error message
              example: Invalid request body
            errors:
              type: array
              description: Detailed validation errors (for VALIDATION_ERROR)
              items:
                type: object
                properties:
                  path:
                    type: string
                    example: url
                  message:
                    type: string
                    example: URL is required
            stack:
              type: string
              description: Stack trace (only in development)
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication (not yet implemented)